Don't send SIGKILL 0 when deleting script that never finished spawning

+++ openc3-cosmos-script-runner-api/app/controllers/running_script_controller.rb
@@ -66,43 +66,47 @@ class RunningScriptController < ApplicationController
     running_script = OpenC3::ScriptStatusModel.get_model(name: params[:id], scope: params[:scope])
     if running_script
       target_name = running_script.filename.split('/')[0]
-      pid = running_script.pid.to_i
+      pid = running_script.pid&.to_i
       return unless authorization('script_run', target_name: target_name)
       running_script_publish("cmd-running-script-channel:#{params[:id]}", "stop")
 
-      # Give the process 1 second to stop from stop message
-      stopped = false
-      start_time = Time.now
-      while Time.now - start_time < 1.0
-        begin
-          Process.getpgid(pid.to_i)
-          sleep 0.1
-        rescue Errno::ESRCH
-          stopped = true
-          break
-        end
-      end
-
-      # If the process is still running
-      # Send a SIGINT and give it one more second
-      if not stopped
-        Process.kill("SIGINT", pid)
+      if pid
+        # Give the process 1 second to stop from stop message
+        stopped = false
         start_time = Time.now
         while Time.now - start_time < 1.0
           begin
-            Process.getpgid(pid.to_i)
+            Process.getpgid(pid)
             sleep 0.1
           rescue Errno::ESRCH
             stopped = true
             break
           end
         end
+
+        # If the process is still running
+        # Send a SIGINT and give it one more second
+        if not stopped
+          Process.kill("SIGINT", pid)
+          start_time = Time.now
+          while Time.now - start_time < 1.0
+            begin
+              Process.getpgid(pid)
+              sleep 0.1
+            rescue Errno::ESRCH
+              stopped = true
+              break
+            end
+          end
+        end
+
+        # If the process is still running send a hard kill signal
+        if not stopped
+          Process.kill("SIGKILL", pid)
+        end
       end
 
-      # If the process is still running send a hard kill signal
-      if not stopped
-        Process.kill("SIGKILL", pid)
-
+      if pid.nil? or not stopped
         # Also need to cleanup the status model in this case because the process will not die and cleanup
         running_script.end_time = Time.now.utc.iso8601
         running_script.state = "killed"
